requirements.txt
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.31.1
openai==1.12.0